% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNOT.R
\name{CNOT}
\alias{CNOT}
\title{CNOT gate}
\usage{
CNOT(n)
}
\description{
This function operates the CNOT gate on a conformable input matrix
}
\examples{
initialize_()
CNOT(I4)
CNOT(Q11)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
