% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hadamard.R
\name{Hadamard}
\alias{Hadamard}
\title{Hadamard Gate}
\usage{
Hadamard(n)
}
\description{
This function operates the Hadamard gate on a conformable input matrix
}
\examples{
initialize_()
Hadamard(Q0)
Hadamard(I2)
Hadamard(Hadamard(Q1))

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
