% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialization}
\usage{
init()
}
\value{
No return value, generates the required variables/parameters.
}
\description{
Builds the parameters in the required environment after initialization
}
\examples{
init()
Q$Q110
Q$Qt12
Q$Q_minus
Q$lambda4

}
\references{
\url{https://arxiv.org/pdf/Quant-ph/0512125.pdf}\cr
\url{https://arxiv.org/pdf/0910.4222.pdf}\cr
\url{https://arxiv.org/pdf/Quant-ph/9703032.pdf}\cr
\url{https://en.wikipedia.org/wiki/Quantum_computing}\cr
\url{https://en.wikipedia.org/wiki/Qubit}\cr
\url{https://en.wikipedia.org/wiki/Qutrit}\cr
\url{https://en.wikipedia.org/wiki/Clebsch\%E2\%80\%93Gordan_coefficients_for_SU(3)}\cr
}
