% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levi_civita.R
\name{levi_civita}
\alias{levi_civita}
\title{Levi-Civita symbol}
\usage{
levi_civita(i, j, k)
}
\description{
This function computes the Levi-Civita symbol depending on the permutations of the three inputs, lying in 0 to 2
}
\examples{
initialize_()
levi_civita(0, 2, 1)
levi_civita(1, 2, 0)
levi_civita(1, 2, 1)

}
\references{
\url{https://en.wikipedia.org/wiki/Levi-Civita_symbol}\cr
}
\keyword{Linear}
\keyword{algebra,}
\keyword{analysis}
\keyword{tensor}
