% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASH.R
\name{NASH}
\alias{NASH}
\title{Nash Equilibrium}
\usage{
NASH(P1, P2)
}
\description{
This function finds out the Nash equilibria of the 2-D payoff matrix for the players. The input parameters are equal dimensional payoff matrices for the first and the second players.
}
\examples{
initialize_()
Alice <- matrix(c(4, 3, 2, 4, 4, 2, 1, 0, 3, 5, 3, 5, 2, 3, 1, 3), ncol=4, byrow=TRUE)
Bob <- matrix(c(0, 2, 3, 8, 2, 1, 2, 2, 6, 5, 1, 0, 3, 2, 2, 3), ncol=4, byrow=TRUE)
NASH(Alice, Bob)

}
\references{
\url{https://arxiv.org/abs/1512.06808}\cr
\url{https://en.wikipedia.org/wiki/Nash_equilibrium}\cr
}
\keyword{Game}
\keyword{Theory,}
\keyword{dominance}
\keyword{equilibrium,}
\keyword{strategy}
\keyword{strict}
