% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PayoffMatrix_QHawkDove.R
\name{PayoffMatrix_QHawkDove}
\alias{PayoffMatrix_QHawkDove}
\title{Quantum Hawk and Dove game: Payoff Matrix}
\usage{
PayoffMatrix_QHawkDove(moves, v, j, D)
}
\arguments{
\item{moves}{a list of matrices}

\item{v}{a number}

\item{j}{a number}

\item{D}{a number}
}
\description{
This function generates the payoff matrix for the Quantum Hawk and Dove game for all the four combinations of 'p' and 'q'. 'moves' is a list of two possible strategies for each of the players and {v, j, D} are the value of resource, cost of injury and cost of displaying.
}
\examples{
init()
moves <- list(I2, sigmaX(I2))
PayoffMatrix_QHawkDove(moves, 50, -100, -10)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0108075.pdf}\cr
}
