% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PayoffMatrix_QPD.R
\name{PayoffMatrix_QPD}
\alias{PayoffMatrix_QPD}
\title{Quantum Prisoner's Dilemma game: Payoff Matrix}
\usage{
PayoffMatrix_QPD(moves, w, x, y, z)
}
\description{
This function generates the payoff matrix for the Quantum Prisoner's Dilemma game . 'moves' is a list of the possible strategies for each of the players and {w, x, y, z} are the payoffs for the players corresponding to the choices available to them with the chain of inequalities, z>w>x>y.
This function also plots the probability distribution plots of the qubits for all the possible combinations of the strategies of the players.
}
\examples{
initialize_()
moves <- list(I2, sigmaX(I2), Hadamard(I2), sigmaZ(I2))
PayoffMatrix_QPD(moves, 3, 1, 0, 5)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0004076.pdf}\cr
}
\keyword{Dilemma}
\keyword{Game}
\keyword{Prisoner's}
\keyword{Quantum}
\keyword{Theory,}
