% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase.R
\name{Phase}
\alias{Phase}
\title{Phase Gate}
\usage{
Phase(n)
}
\description{
This function operates the Phase gate on a conformable input matrix
}
\examples{
initialize_()
Phase(I2)
Phase(Q_plus)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
