% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhaseDagger.R
\name{PhaseDagger}
\alias{PhaseDagger}
\title{Hermitian Transpose of the Phase Gate}
\usage{
PhaseDagger(n)
}
\arguments{
\item{n}{a vector/matrix}
}
\description{
This function operates the hermitian transpose of the Phase gate on a conformable input matrix
}
\examples{
init()
Conj(t(Phase(I2)))==PhaseDagger(I2)
PhaseDagger(Q_plus)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
