% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDuels_Bob_payoffs.R
\name{QDuels_Bob_payoffs}
\alias{QDuels_Bob_payoffs}
\title{Quantum Two Person Duel game}
\usage{
QDuels_Bob_payoffs(Psi, n, a, b, alpha1, alpha2, beta1, beta2)
}
\arguments{
\item{Psi}{a vector representing the initial quantum state}

\item{n}{an integer}

\item{a}{a number}

\item{b}{a number}

\item{alpha1}{a number}

\item{alpha2}{a number}

\item{beta1}{a number}

\item{beta2}{a number}
}
\description{
This function returns the expected payoff to Bob for three possible cases for the Quantum Duel game:
\enumerate{
\item The game is continued for 'n' rounds and none of the players shoots at the air.
\item The game is continued for 2 rounds and Alice shoots at the air in her second round.
\item The game is continued for 2 rounds and Bob shoots at the air in her second round.
Psi is the initial state of the quantum game, 'n' is the number of rounds, 'a' is the probability of Alice missing the target, 'b' is the probability of Bob missing the target, and
{'alpha1', 'alpha2', 'beta1', 'beta2'} are arbitrary phase factors that lie in -pi to pi that control the outcome of a poorly performing player
}
}
\examples{
init()
QDuels_Bob_payoffs(Q11, 5, 0.666666, 0.5, 0, 0, 0.2, 0.7)
Qs <- (Q0+Q1)/sqrt(2)
Psi <- kronecker(Qs, Qs)
QDuels_Bob_payoffs(Psi, 5, 0.666666, 0.5, 0, 0, 0.2, 0.7)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0305058.pdf}\cr
}
