% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDuelsPlot4.R
\name{QDuelsPlot4}
\alias{QDuelsPlot4}
\title{Quantum Two Person Duel game}
\usage{
QDuelsPlot4(Psi, alpha1, alpha2)
}
\description{
This function helps us to plot the improvement in Bob's expected payoff as a function of 'a' and 'b', if Bob chooses to fire at the air in her second shot, in a two round game. Psi is the initial state of the quantum game, 'n' is the number of rounds, 'a' is the probability of Alice missing the target, 'b' is the probability of Bob missing the target, and
{'alpha1', 'alpha2', 'beta1', 'beta2'} are arbitrary phase factors that lie in \link{-pi, pi} that control the outcome of a poorly performing player.
}
\examples{
initialize_()
Qs <- (Q0+Q1)/sqrt(2)
Psi <- kronecker(Q1, Qs)
QDuelsPlot4(Psi, pi/3, pi/6)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0305058.pdf}\cr
}
\keyword{Duel}
\keyword{Game}
\keyword{Person}
\keyword{Quantum}
\keyword{Theory,}
\keyword{Two}
