% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMeasure.R
\name{QMeasure}
\alias{QMeasure}
\title{Measurement}
\usage{
QMeasure(n)
}
\arguments{
\item{n}{a vector representing a quantum state}
}
\description{
This function performs a projective measurement of a quantum state 'n', in the computational basis and plots the corresponding probability distributions of the qubits.
}
\examples{
init()
QMeasure(Q11010001)

}
\references{
\url{https://books.google.co.in/books?id=66TgFp2YqrAC&pg=PA25&redir_esc=y}\cr
\url{https://en.wikipedia.org/wiki/Measurement_in_quantum_mechanics}\cr
}
