% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMontyHall.R
\name{QMontyHall}
\alias{QMontyHall}
\title{Quantum Monty Hall Problem}
\usage{
QNewcomb(Psi_in, gamma, Ahat, Bhat)
}
\description{
This function simulates the quantum version of the Monty Hall problem, by taking in 'Psi_in' as the initial quantum state of the game, 'gamma' lying in 0 to pi/2, Ahat and Bhat as the choice operators in SU(3) for Alice and Bob respectively as the inputs. It returns the expected payoffs to Alice and Bob after the end of the game.
}
\examples{
initialize_()
Psi_in <- kronecker(Qt0, (Qt00+Qt11+Qt22)/sqrt(3))
QMontyHall(Psi_in, pi/4, Identity3, Hhat)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0109035.pdf}\cr
}
\keyword{Game}
\keyword{Hall}
\keyword{Monty}
\keyword{Problem}
\keyword{Quantum}
\keyword{Theory,}
