% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPennyFlip.R
\name{QPennyFlip}
\alias{QPennyFlip}
\title{Quantum Penny Flip game}
\usage{
QPennyFlip(initial_state, strategies_Alice, strategies_Bob)
}
\description{
This function simulates the Quantum Penny Flip game by taking in the initial state of the game that is set by Alice and the strategies available to Alice and Bob. It returns the final state of the game along with the probability distribution of the qubits after measurement of the final state.
}
\examples{
initialize_()
psi <- (u+d)/sqrt(2)
S1 <- sigmaX(I2)
S2 <- I2
H <- Hadamard(I2)
SA <- list(S1, S2)
SB <- list(H)
QPennyFlip(psi, SA,SB)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/9804010.pdf}\cr
}
\keyword{Flip}
\keyword{Game}
\keyword{Penny}
\keyword{Quantum}
\keyword{Theory,}
