% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_count.R
\name{row_count}
\alias{row_count}
\title{Number of rows of a vector/matrix}
\usage{
row_count(M)
}
\description{
This function counts the number of rows of a vector or a matrix
}
\examples{
initialize_()
row_count(Q01)
row_count(lambda5)
row_count(Qt12)

}
\keyword{Matrix}
\keyword{Vector,}
