% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ry.R
\name{Ry}
\alias{Ry}
\title{Rotation operation about y-axis of the Bloch sphere}
\usage{
Ry(n, theta)
}
\description{
This function operates the Rotation gate about the y-axis of the Bloch sphere by an angle 'theta' on a conformable input matrix 'n'.
}
\examples{
initialize_()
Ry(Q1, pi/3)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
\url{http://www.physics.udel.edu/~msafrono/650/Lecture\%204\%20-\%205.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
