% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaZ.R
\name{sigmaZ}
\alias{sigmaZ}
\title{Pauli-Z gate}
\usage{
sigmaZ(n)
}
\description{
This function operates the Pauli-Z gate on a conformable input matrix.
}
\examples{
initialize_()
sigmaZ(I2)
sigmaZ(Hadamard(I2))
sigmaZ(Q0)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
\url{http://www.physics.udel.edu/~msafrono/650/Lecture\%204\%20-\%205.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
