% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWAP.R
\name{SWAP}
\alias{SWAP}
\title{SWAP gate}
\usage{
SWAP(n)
}
\description{
This function operates the SWAP gate on a conformable input matrix.
}
\examples{
initialize_()
SWAP(I4)
SWAP(Q10)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
