% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.R
\name{T}
\alias{T}
\title{T gate}
\usage{
T(n)
}
\description{
This function operates the T gate on a conformable input matrix.
}
\examples{
initialize_()
T(I2)
T(Q_minus)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
