% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Walsh.R
\name{Walsh}
\alias{Walsh}
\title{Walsh-Hadamard gate}
\usage{
Walsh(n)
}
\description{
This function operates the Walsh-Hadamard gate on a conformable input matrix.
}
\examples{
initialize_()
Walsh(I2)
Walsh(Q0)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://en.wikipedia.org/wiki/Hadamard_transform}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
