% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Walsh16.R
\name{Walsh16}
\alias{Walsh16}
\title{Walsh-Hadamard gate}
\usage{
Walsh16(n)
}
\description{
This function operates the Walsh-16 gate on a conformable input matrix.
}
\examples{
initialize_()
Walsh16(I16)
Walsh16(Q1001)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://en.wikipedia.org/wiki/Hadamard_transform}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
