% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bell.R
\name{Bell}
\alias{Bell}
\title{Bell States}
\usage{
Bell(qubit1, qubit2)
}
\description{
The function builds one of the four Bell states, according to the input qubits
}
\examples{
initialize_()
Bell(0, 0)
Bell(0, 1)
Bell(1, 0)
Bell(1, 1)

}
\references{
\url{https://en.wikipedia.org/wiki/Bell_state}\cr
\url{https://books.google.co.in/books?id=66TgFp2YqrAC&pg=PA25&redir_esc=y}\cr
}
\keyword{Bell}
\keyword{Quantum}
\keyword{basis}
\keyword{entanglement,}
