% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bell.R
\name{Bell}
\alias{Bell}
\title{Bell States}
\usage{
Bell(qubit1, qubit2)
}
\arguments{
\item{qubit1}{1st input qubit}

\item{qubit2}{2nd input qubit}
}
\value{
One of the Bell states as a vector depending on the input qubits.
}
\description{
The function builds one of the four Bell states, according to the input qubits
}
\examples{
init()
Bell(Q$Q0, Q$Q0)
Bell(Q$Q0, Q$Q1)
Bell(Q$Q1, Q$Q0)
Bell(Q$Q1, Q$Q1)

}
\references{
\url{https://en.wikipedia.org/wiki/Bell_state}\cr
\url{https://books.google.co.in/books?id=66TgFp2YqrAC&pg=PA25&redir_esc=y}\cr
}
