% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNOT.R
\name{CNOT}
\alias{CNOT}
\title{CNOT gate}
\usage{
CNOT(n)
}
\arguments{
\item{n}{A vector/matrix}
}
\value{
A matrix or a vector after performing the CNOT gate operation on a conformable input matrix or a vector.
}
\description{
This function operates the CNOT gate on a conformable input matrix/vector
}
\examples{
init()
CNOT(Q$I4)
CNOT(Q$Q11)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
