% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_count.R
\name{col_count}
\alias{col_count}
\title{Number of columns of a vector/matrix}
\usage{
col_count(M)
}
\arguments{
\item{M}{a A vector/matrix}
}
\description{
This function counts the number of columns of a vector or a matrix
}
\examples{
init()
col_count(Q11)
col_count(lambda4)
col_count(I2)

}
