% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_count.R
\name{col_count}
\alias{col_count}
\title{Number of columns of a vector/matrix}
\usage{
col_count(M)
}
\arguments{
\item{M}{A vector/matrix}
}
\value{
An integer that gives the number of columns in a vector or a matrix.
}
\description{
This function counts the number of columns of a vector or a matrix
}
\examples{
init()
col_count(Q$Q11)
col_count(Q$lambda4)
col_count(Q$I2)

}
