% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fredkin.R
\name{Fredkin}
\alias{Fredkin}
\title{Fredkin Gate}
\usage{
Fredkin(n)
}
\description{
This function operates the Fredkin gate on a conformable input matrix
}
\examples{
initialize_()
Fredkin(I8)
Fredkin(Q110)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
