% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fredkin.R
\name{Fredkin}
\alias{Fredkin}
\title{Fredkin Gate}
\usage{
Fredkin(n)
}
\arguments{
\item{n}{A vector/matrix}
}
\value{
A matrix or a vector after performing the Fredkin gate operation on a conformable input matrix or a vector.
}
\description{
This function operates the Fredkin gate on a conformable input matrix/vector
}
\examples{
init()
Fredkin(Q$I8)
Fredkin(Q$Q110)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
