% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hadamard.R, R/Ising.R
\name{Hadamard}
\alias{Hadamard}
\title{Hadamard Gate}
\usage{
Hadamard(n)

Ising(n, phi)
}
\description{
This function operates the Hadamard gate on a conformable input matrix

This function operates the Ising gate on a conformable input matrix. A phase (phi) is also given as another input.
}
\examples{
initialize_()
Hadamard(Q0)
Hadamard(I2)
Hadamard(Hadamard(Q1))

initialize_()
Ising(I4, pi/4)
Ising(Q11, pi/6)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr

\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
