% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levi_civita.R
\name{levi_civita}
\alias{levi_civita}
\title{Levi-Civita symbol}
\usage{
levi_civita(i, j, k)
}
\arguments{
\item{i}{an integer 0, 1 or 2}

\item{j}{an integer 0, 1 or 2}

\item{k}{an integer 0, 1 or 2}
}
\value{
0, 1 or -1 after computing the Levi-Civita symbol depending on the permutations of the three inputs 0, 1 and 2
}
\description{
This function computes the Levi-Civita symbol depending on the permutations of the three inputs, lying in 0 to 2
}
\examples{
init()
levi_civita(0, 2, 1)
levi_civita(1, 2, 0)
levi_civita(1, 2, 1)

}
\references{
\url{https://en.wikipedia.org/wiki/Levi-Civita_symbol}\cr
}
