% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDuelsPlot1.R
\name{QDuelsPlot1}
\alias{QDuelsPlot1}
\title{Quantum Two Person Duel game}
\usage{
QDuelsPlot1(Psi, n, a, b, beta1, beta2)
}
\description{
This function helps us to plot Alice's and Bob's expected payoffs as functions of 'alpha1' and 'alpha2'. Psi is the initial state of the quantum game, 'n' is the number of rounds, 'a' is the probability of Alice missing the target, 'b' is the probability of Bob missing the target, and
{'alpha1', 'alpha2', 'beta1', 'beta2'} are arbitrary phase factors that lie in -pi to pi that control the outcome of a poorly performing player.
}
\examples{
initialize_()
QDuelsPlot1(Q10, 2, 0.66666, 0.5, 0.2, 0.8)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0305058.pdf}\cr
}
\keyword{Duel}
\keyword{Game}
\keyword{Person}
\keyword{Quantum}
\keyword{Theory,}
\keyword{Two}
