% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFT.R
\name{QFT}
\alias{QFT}
\title{Quantum Fourier Transform}
\usage{
QFT(y)
}
\arguments{
\item{y}{an integer}
}
\value{
A vector representing the Quantum Fourier transformation of the state \code{|y>} from the computational basis to the Fourier basis.
}
\description{
This function performs Quantum Fourier Transform for a given state \code{|y>} from the computational basis to the Fourier basis.
}
\examples{
init()
QFT(5)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://books.google.co.in/books?id=66TgFp2YqrAC&pg=PA25&redir_esc=y}\cr
\url{https://en.wikipedia.org/wiki/Quantum_Fourier_transform}\cr
}
