% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QHawkDove.R
\name{QHawkDove}
\alias{QHawkDove}
\title{Quantum Hawk and Dove game}
\usage{
QHawkDove(p, q, moves, v, j, D)
}
\description{
This function returns the expected payoffs to Alice and Bob with respect to the probabilities 'p' and 'q'. p+q should equal 1 and 'moves' is a list of two possible strategies for each of the players and {v, j, D} are the value of resource, cost of injury and cost of displaying.
}
\examples{
initialize_()
moves <- list(I2, sigmaX(I2))
QHawkDove(0, 1, moves, 50, -100, -10)
QHawkDove(0, 0, moves, 50, -100, -10)

}
\references{
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0208069.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0108075.pdf}\cr
}
\keyword{Dove}
\keyword{Game}
\keyword{Hawk}
\keyword{Quantum}
\keyword{Theory,}
\keyword{and}
