% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_count.R
\name{row_count}
\alias{row_count}
\title{Number of rows of a vector/matrix}
\usage{
row_count(M)
}
\arguments{
\item{M}{A vector/matrix}
}
\value{
An integer that gives the number of rows in a vector or a matrix.
}
\description{
This function counts the number of rows of a vector or a matrix
}
\examples{
init()
row_count(Q$Q01)
row_count(Q$lambda5)
row_count(Q$Qt12)

}
