% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaY.R
\name{sigmaY}
\alias{sigmaY}
\title{Pauli-Y gate}
\usage{
sigmaY(n)
}
\description{
This function operates the Pauli-X gate on a conformable input matrix.
}
\examples{
initialize_()
sigmaY(I2)
sigmaY(Hadamard(I2))
sigmaY(Q0)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
\url{http://www.physics.udel.edu/~msafrono/650/Lecture\%204\%20-\%205.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
