% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaZ.R
\name{sigmaZ}
\alias{sigmaZ}
\title{Pauli-Z gate}
\usage{
sigmaZ(n)
}
\arguments{
\item{n}{a vector/matrix}
}
\value{
A matrix or a vector after performing the Pauli-Z gate operation on a conformable input matrix or a vector.
}
\description{
This function operates the Pauli-Z gate on a conformable input matrix or a vector.
}
\examples{
init()
sigmaZ(Q$I2)
sigmaZ(Hadamard(Q$I2))
sigmaZ(Q$Q0)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
\url{http://www.physics.udel.edu/~msafrono/650/Lecture\%204\%20-\%205.pdf}\cr
}
