% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWAP.R
\name{SWAP}
\alias{SWAP}
\title{SWAP gate}
\usage{
SWAP(n)
}
\arguments{
\item{n}{a vector/matrix}
}
\value{
A matrix or a vector after performing the SWAP gate operation on a conformable input matrix or a vector.
}
\description{
This function operates the SWAP gate on a conformable input matrix or a vector.
}
\examples{
init()
SWAP(Q$I4)
SWAP(Q$Q10)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
