% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDagger.R
\name{TDagger}
\alias{TDagger}
\title{Hermitian Transpose of the T gate}
\usage{
TDagger(n)
}
\arguments{
\item{n}{a vector/matrix}
}
\value{
A matrix or a vector after performing the operation of the hermitian transpose of the T gate on a conformable input matrix or a vector.
}
\description{
This function operates the hermitian transpose of the T gate on a conformable input matrix or a vector.
}
\examples{
init()
TDagger(Q$I2)
TDagger(Q$Q_plus)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
