% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toffoli.R
\name{Toffoli}
\alias{Toffoli}
\title{Toffoli gate}
\usage{
Toffoli(n)
}
\description{
This function operates the Toffoli gate on a conformable input matrix.
}
\examples{
initialize_()
Toffoli(I8)
Toffoli(Q010)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
\keyword{Quantum}
\keyword{gates}
\keyword{logic}
\keyword{operations,}
\keyword{quantum}
