% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toffoli.R
\name{Toffoli}
\alias{Toffoli}
\title{Toffoli gate}
\usage{
Toffoli(n)
}
\arguments{
\item{n}{a vector/matrix}
}
\value{
A matrix or a vector after performing the Toffoli gate operation on a conformable input matrix or a vector.
}
\description{
This function operates the Toffoli gate on a conformable input matrix or a vector.
}
\examples{
init()
Toffoli(Q$I8)
Toffoli(Q$Q010)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
}
