% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Walsh4.R
\name{Walsh4}
\alias{Walsh4}
\title{Walsh-Hadamard gate}
\usage{
Walsh4(n)
}
\arguments{
\item{n}{a vector/matrix}
}
\value{
A matrix or a vector after performing the Walsh-4 gate operation on a conformable input matrix or a vector.
}
\description{
This function operates the Walsh-4 gate on a conformable input matrix or a vector.
}
\examples{
init()
Walsh4(Q$I4)
Walsh4(Q$Q10)

}
\references{
\url{https://en.wikipedia.org/wiki/Quantum_logic_gate}\cr
\url{http://www2.optics.rochester.edu/~stroud/presentations/muthukrishnan991/LogicGates.pdf}\cr
\url{https://arxiv.org/pdf/quant-ph/0506219.pdf}\cr
\url{https://en.wikipedia.org/wiki/Hadamard_transform}\cr
}
